airflow/dags/elt/opralog/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --no-annotate --output-file=./requirements.txt --strip-extras ./requirements.in
#
aiobotocore==2.15.2
aiohappyeyeballs==2.4.3
aiohttp==3.11.5
aioitertools==0.12.0
aiosignal==1.3.1
attrs==24.2.0
botocore==1.35.36
connectorx==0.4.0
frozenlist==1.5.0
fsspec==2024.10.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
polars==1.13.1
propcache==0.2.0
pyarrow==18.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3fs==2024.10.0
six==1.16.0
urllib3==2.2.3
wrapt==1.16.0
yarl==1.17.2
